Only download coveralls on CI
This makes the coveralls check optional when normalling building. It
removes the following lines from being printed when compiling projects
that depend on it:

```
===> Fetching coveralls (from
{git,"https://github.com/markusn/coveralls-erl",
                     {ref,"57a190502cd2606d2b5cc142bd5cec1ec7c497b2"}})
===> Compiling coveralls
_build/default/plugins/coveralls/src/coveralls.erl:103: Warning:
random:uniform/1: the 'random' module is deprecated; use the 'rand'
module instead

```

It also bumps coveralls to use the version stored on hex since the
specified version seemed pretty old.